Remove default local AI context override
